chore: move components related to file upload under FileUpload directory and generate test files for those components
